Move the tenant ID converter registry into a separate class.
