+++ gcd/src/main/java/org/spine3/server/storage/datastore/DatastoreStorageFactory.java
@@ -33,10 +33,10 @@ import org.spine3.server.projection.ProjectionStorage;
 import org.spine3.server.stand.StandStorage;
 import org.spine3.server.storage.RecordStorage;
 import org.spine3.server.storage.StorageFactory;
-import org.spine3.server.storage.datastore.tenant.DatastoreTenants;
 import org.spine3.server.storage.datastore.tenant.Namespace;
 import org.spine3.server.storage.datastore.tenant.NamespaceSupplier;
 import org.spine3.server.storage.datastore.tenant.NamespaceToTenantIdConverter;
+import org.spine3.server.storage.datastore.tenant.TenantConverterRegistry;
 import org.spine3.server.storage.datastore.type.DatastoreColumnType;
 import org.spine3.server.storage.datastore.type.DatastoreTypeRegistryFactory;
 import org.spine3.type.TypeUrl;
@@ -322,8 +322,8 @@ public class DatastoreStorageFactory implements StorageFactory {
             }
             if (namespaceToTenantIdConverter != null) {
                 final ProjectId projectId = ProjectId.of(datastore);
-                DatastoreTenants.registerNamespaceConverter(projectId,
-                                                            namespaceToTenantIdConverter);
+                TenantConverterRegistry.registerNamespaceConverter(projectId,
+                                                                   namespaceToTenantIdConverter);
             }
 
             return new DatastoreStorageFactory(this);
